refactor: playerのディレクトリ変更
