+++ src/components/molecules/left-controller.component.ts
@@ -1,5 +1,5 @@
 import { BaseComponent } from '../common/base.component'
-import { Player } from '../../entities/player'
+import { Player } from '../../entities/player/player'
 const sheet = new CSSStyleSheet()
 sheet.replaceSync(`
   .button-container {
